[DI-304] Fix search for template usage
The "temple ls user template" usage seemed to be broken,
it would return incomplete, inconsistent, or non-existent results.

I found that using the filename: term I could search within an
explicitly named file instead of combining the in:path and in:file
directives to try to do the same thing. I also found that including
double-quotes in the query string made it less reliable,
and according to the search documentation they are ignored anyway
(see https://help.github.com/en/articles/searching-code#considerations-for-code-search).

I removed the double-quotes from the query string and switched to
using the filename: search and results of my local testing have
been reliable and (apparently) complete.